+++ docs/src/component-blocks/aside.tsx
@@ -1,0 +1,25 @@
+import { fields, component } from '@keystatic/core';
+
+import { preview } from './aside.preview';
+
+export const schema = {
+  icon: fields.text({
+    label: 'Emoji icon...',
+  }),
+  content: fields.child({
+    kind: 'block',
+    placeholder: 'Aside...',
+    formatting: {
+      inlineMarks: 'inherit',
+      softBreaks: 'inherit',
+      listTypes: 'inherit',
+    },
+    links: 'inherit',
+  }),
+};
+
+export const aside = component({
+  label: 'Aside',
+  schema,
+  preview,
+});